upd: Reorganiza as tools do fluxo de codificação
