+++ adk/shared/tools/filesystem.py
@@ -2,16 +2,49 @@
 
 As tools exclusivas do fluxo de codificação (tool_criar_arquivo,
 tool_ler_arquivo, tool_substituir_trecho, tool_salvar_relatorio) vivem em
-shared/tools/coding_review/filesystem.py.
+shared/tools/coding_tools/filesystem_coding.py.
 """
 
 import re
 from pathlib import Path
 from typing import Optional
 
-from shared.tools._path_utils import _resolver_caminho
-
 ID_REQ_PATTERN = re.compile(r"^[A-Z]{1,4}-\d{3}$")
+
+
+def _resolver_caminho(caminho: str, base_dir: Optional[str] = None) -> Path:
+    """Resolve caminho relativo ao base_dir (se informado) com proteção anti-traversal.
+
+    Args:
+        caminho: Caminho informado pelo agente.
+        base_dir: Diretório base do agente no workspace (opcional).
+
+    Returns:
+        Path resolvido e validado.
+
+    Raises:
+        ValueError: Se o caminho tenta escapar do base_dir (absolute ou ..).
+    """
+    if base_dir is None:
+        return Path(caminho)
+
+    base = Path(base_dir).resolve()
+    rel = Path(caminho)
+
+    if rel.is_absolute():
+        raise ValueError(
+            f"Caminho absoluto não permitido com base_dir: '{caminho}'. "
+            f"Use caminhos relativos ao seu diretório de trabalho."
+        )
+
+    if ".." in rel.parts:
+        raise ValueError(
+            f"Path traversal não permitido: '{caminho}'. "
+            f"Não use '..' no caminho."
+        )
+
+    return base / rel
+
 
 # Layout legado: paths absolutos baseados no CWD.
 _PASTAS_LEGADO = {
