+++ adk/src/agents/context_engineer/agent.py
@@ -5,7 +5,7 @@ composta com workspace binding — tool_salvar_task escreve em workspace/tasks/.
 """
 
 from shared.agent_factory import create_se_agent
-from shared.tools.coding_review.context_engineer_task import tool_salvar_task_adk
+from shared.tools.coding_tools.context_engineer_tools import tool_salvar_task_adk
 
 from . import prompt, schemas
