+++ adk/tests/unit/test_filesystem_base_dir.py
@@ -3,7 +3,7 @@
 import pytest
 from pathlib import Path
 
-from AI4ES.adk.shared.tools.coding_review.filesystem_coding import (
+from shared.tools.coding_tools.filesystem_coding import (
     tool_criar_arquivo,
     tool_ler_arquivo,
     tool_substituir_trecho,
@@ -12,8 +12,8 @@ from AI4ES.adk.shared.tools.coding_review.filesystem_coding import (
 from shared.tools.filesystem import (
     tool_ler_workspace,
     tool_listar_workspace,
+    _resolver_caminho,
 )
-from shared.tools._path_utils import _resolver_caminho
 
 
 def test_criar_arquivo_sem_base_dir_funciona(tmp_path, monkeypatch):
