+++ adk/tests/unit/test_filesystem_tools.py
@@ -13,7 +13,7 @@ from pathlib import Path
 import pytest
 
 sys.path.insert(0, str(Path(__file__).parent))
-from AI4ES.adk.shared.tools.coding_review.filesystem_coding import tool_criar_arquivo
+from shared.tools.coding_tools.filesystem_coding import tool_criar_arquivo
 from shared.tools.filesystem import tool_salvar_artefato_requisito
